==> 1. Mark Duplicative Complaints.ipynb <==
import pandas as pd
from datetime import datetime
from difflib import SequenceMatcher
import numpy as np
from fuzzywuzzy import fuzz

from IPython.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

# Define the file path relative to the current notebook's location
file_path = "../data/complaints 5.9.2023.csv"

# Read the CSV file into a DataFrame
cfpb_df = pd.read_csv(file_path)

cfpb_df.sample(5)

%%time
# Convert the 'date' column to datetime format
print("Before droping nan narrative: ", len(cfpb_df))
cfpb_df.dropna(subset=['Consumer complaint narrative'], inplace=True)
cfpb_df['Date received'] = pd.to_datetime(cfpb_df['Date received'])
cfpb_df['narr_len'] = cfpb_df['Consumer complaint narrative'].apply(lambda x:len(str(x)))
cfpb_df['days_to_today'] = (datetime.now().date() - cfpb_df['Date received'].dt.date).dt.days

cfpb_df['narr_len'] = cfpb_df['narr_len'].astype(int)
cfpb_df['days_to_today'] = cfpb_df['days_to_today'].astype(int)


cfpb_df['narr_head'] = cfpb_df['Consumer complaint narrative'].apply(lambda x: x[:200])
print("After droping nan narrative: ", len(cfpb_df))

small_cfpb_df = cfpb_df[['Product', 'Issue', 'State', 'ZIP code','Complaint ID','narr_len', 'days_to_today', 'narr_head']].copy()
small_cfpb_df[['Product', 'Issue', 'State', 'ZIP code']] = small_cfpb_df[['Product', 'Issue', 'State', 'ZIP code']].fillna('')

# when running group by, the input is a mini dataframe with same ['Product', 'Issue', 'State', 'ZIP code']
def find_duplicate_narr(df):
    small_df = df[['Complaint ID','narr_len', 'days_to_today', 'narr_head']].copy()
    
    def find_dupi_in_small_df(row_narr_len, row_to_day, row_short_narr, small_df):
#         tmp_df = small_df.query("'narr_len'<=%s & 'narr_len'>=%s & 'days_to_today'<= %s & 'days_to_today'>= %s" % (int(row_narr_len*1.2),
#                                                                                                                    int(row_narr_len*0.8),
#                                                                                                                    int(row_to_day+15),
#                                                                                                                    int(row_to_day-15))).copy()
        tmp_df = small_df.query("narr_len <= @row_narr_len*1.2 & narr_len >= @row_narr_len*0.8 & days_to_today <= @row_to_day+15 & days_to_today >= @row_to_day-15").copy()    
        # tmp_df['similarity_to_row'] = tmp_df['narr_head'].apply(lambda x: SequenceMatcher(None, x, row_short_narr).ratio())
        tmp_df['similarity_to_row'] = tmp_df['narr_head'].apply(lambda x: fuzz.WRatio(x, row_short_narr)/100)
        
        dupli_df = tmp_df[tmp_df['similarity_to_row']>0.85]
        dupli_id_list = sorted(dupli_df['Complaint ID'].to_list())
#         if len(dupli_id_list)>1:
#             print(dupli_id_list)
        return dupli_id_list
    
    df['dupi_id'] = small_df.apply(lambda row: find_dupi_in_small_df(row['narr_len'], int(row['days_to_today']), row['narr_head'], small_df),axis=1)
    
    return df

%%time
small_cfpb_df = small_cfpb_df.groupby(['Product', 'Issue', 'State', 'ZIP code']).apply(func=find_duplicate_narr)

small_cfpb_df['dupi_len'] = small_cfpb_df['dupi_id'].apply(lambda x: len(x))
small_cfpb_df['dupi_id'] = small_cfpb_df['dupi_id'].apply(lambda x: ";".join([str(y) for y in x]))

merged_df = cfpb_df.merge(small_cfpb_df[['Complaint ID', 'dupi_id', 'dupi_len']], on='Complaint ID', how='left').drop(['narr_head'], axis=1)

merged_df.to_csv('CFPB with Duplicate Marked.csv')
merged_df.to_csv('../data/CFPB with Duplicate Marked fuzz.csv')

merged_df[merged_df.dupi_len>1]

merged_df.shape

349816/1300361
