==> 3.TEST SVM.ipynb <==
import pandas as pd
import numpy as np
import pickle
from tqdm import tqdm

import seaborn as sns
import matplotlib.pyplot as plt

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_selection import SelectKBest, chi2
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report
from imblearn.over_sampling import SMOTE, RandomOverSampler
from sklearn import naive_bayes #import GaussianNB, MultinomialNB
from sklearn.linear_model import SGDClassifier # simulate the behavior of logistic regression using SGDClassifier(loss='log')
from sklearn.metrics import accuracy_score,balanced_accuracy_score,average_precision_score, classification_report

from sklearn.utils import shuffle
from collections import Counter

from IPython.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

train_df = pd.read_csv('cfpb_train.csv')
test_df = pd.read_csv('cfpb_test.csv')
dev_df = pd.read_csv('cfpb_dev.csv')

train_df['Consumer complaint narrative'] = train_df['Consumer complaint narrative'].fillna('').astype(str)
test_df['Consumer complaint narrative'] = test_df['Consumer complaint narrative'].fillna('').astype(str)
dev_df['Consumer complaint narrative'] = dev_df['Consumer complaint narrative'].fillna('').astype(str)

train_df['debt_collection'] = (train_df['Product'] == 'Debt collection').astype(int)
test_df['debt_collection'] = (test_df['Product'] == 'Debt collection').astype(int)
dev_df['debt_collection'] = (dev_df['Product'] == 'Debt collection').astype(int)

train_df_sample = train_df.sample(10000)

# Load the trained vectorizer
with open('tfidf_vectorizer_train_split_33k.pkl', 'rb') as f:
    loaded_vectorizer = pickle.load(f)
len(loaded_vectorizer.get_feature_names_out())

%%time
# Vectorize the text data with pre-tuned vectorizer
X_train = loaded_vectorizer.transform(train_df['Consumer complaint narrative'])
y_train = train_df['debt_collection']

X_train_sample = loaded_vectorizer.transform(train_df_sample['Consumer complaint narrative'])
y_train_sample = train_df_sample['debt_collection']

X_dev = loaded_vectorizer.transform(dev_df['Consumer complaint narrative'])
y_dev = dev_df['debt_collection']

X_test = loaded_vectorizer.transform(test_df['Consumer complaint narrative'])
y_test = test_df['debt_collection']

# %%time
# selector = SelectKBest(chi2, k=30000)
# X_train = selector.fit_transform(X_train, y_train)
# X_train_sample = selector.transform(X_train_sample)
# X_dev = selector.transform(X_dev)
# X_test = selector.transform(X_test)

%%time
# some balancing
# from imblearn.over_sampling import SMOTE 
# sm = SMOTE(random_state=42)
# X_train_res, y_train_res = sm.fit_resample(X_train, y_train)
# # Shuffle your data
# X_train_res, y_train_res = shuffle(X_train_res, y_train_res)
# define oversampling strategy
oversample = RandomOverSampler(sampling_strategy='minority')
# fit and apply the transform
X_train_res, y_train_res = oversample.fit_resample(X_train, y_train)

# Shuffle your data
X_train_res, y_train_res = shuffle(X_train_res, y_train_res)

### SGD Linear SVM

# Initialize a LogisticRegression model, ‘log_loss’ gives logistic regression, a probabilistic classifier.
clf= SGDClassifier(loss='hinge',random_state=42, alpha=1e-4, n_iter_no_change=3, early_stopping=False)

# The partial fit if you ran out of RAM
batch_size = 50000
n_batches = X_train_res.shape[0] // batch_size

# partial fitting
for i in tqdm(range(n_batches)):
    start = i * batch_size
    end = (i + 1) * batch_size
    batch_X = X_train_res[start:end]
    batch_y = y_train_res[start:end]
    if i == 0:
        clf.partial_fit(batch_X.A, batch_y, classes=np.unique(y_train_res)) 
        # The `.A` here converts the sparse matrix to a dense matrix.
        # This is necessary because GaussianNB doesn't support sparse matrices.
        # We also specify the classes parameter in the first call to partial_fit.
    else:
        clf.partial_fit(batch_X.A, batch_y)

# get the results
y_pred = clf.predict(X_train_sample.toarray())
# y_pred_proba = clf.predict_proba(X_train_sample.toarray())

# Convert lists to arrays for further use
y_pred = np.array(y_pred)
# y_pred_proba = np.array(y_pred_proba)

print("Train")
print("accuracy_score",accuracy_score(y_train_sample, y_pred))
print("balanced_accuracy_score",balanced_accuracy_score(y_train_sample, y_pred))
print("average_precision_score",average_precision_score(y_train_sample, y_pred))
print(classification_report(y_train_sample, y_pred))

# c = Counter(y_pred)
# print("Prediction", c.most_common(2))
# c = Counter(y_train_sample)
# print("Ground Truth",c.most_common(2))

# prediction = pd.DataFrame(y_pred_proba)
# prediction['result'] = y_pred

# df = prediction.copy()
# df.columns = ['neg', 'pos', 'class']
# df['true'] = y_train_sample

# fig, axs = plt.subplots(ncols=2, figsize=(15, 5)) # Adjust figsize as needed

# # Plot Probability Density for Different True Classes
# for class_label in df['true'].unique():
#     sns.kdeplot(df[df['true'] == class_label]['pos'], label=class_label, ax=axs[0])

# axs[0].set_title('Probability Density for Different True Classes')
# axs[0].set_xlabel('Probability')
# axs[0].set_ylabel('Density')
# axs[0].legend(title='True Class')

# # Plot Probability Density for Different Predicted Classes
# for class_label in df['class'].unique():
#     sns.kdeplot(df[df['class'] == class_label]['pos'], label=class_label, ax=axs[1])

# axs[1].set_title('Probability Density for Different Predicted Classes')
# axs[1].set_xlabel('Probability')
# axs[1].set_ylabel('Density')
# axs[1].legend(title='Predicted Class')

# plt.tight_layout()
# plt.show()

# get the results
y_pred = clf.predict(X_dev.toarray())
# y_pred_proba = clf.predict_proba(X_dev.toarray())

# Convert lists to arrays for further use
y_pred = np.array(y_pred)
# y_pred_proba = np.array(y_pred_proba)

print("Train")
print("accuracy_score",accuracy_score(y_dev, y_pred))
print("balanced_accuracy_score",balanced_accuracy_score(y_dev, y_pred))
print("average_precision_score",average_precision_score(y_dev, y_pred))
print(classification_report(y_dev, y_pred))

# c = Counter(y_pred)
# print("Prediction", c.most_common(2))
# c = Counter(y_dev)
# print("Ground Truth",c.most_common(2))

# prediction = pd.DataFrame(y_pred_proba)
# prediction['result'] = y_pred

# df = prediction.copy()
# df.columns = ['neg', 'pos', 'class']
# df['true'] = y_dev

# fig, axs = plt.subplots(ncols=2, figsize=(15, 5)) # Adjust figsize as needed

# # Plot Probability Density for Different True Classes
# for class_label in df['true'].unique():
#     sns.kdeplot(df[df['true'] == class_label]['pos'], label=class_label, ax=axs[0])

# axs[0].set_title('Probability Density for Different True Classes')
# axs[0].set_xlabel('Probability')
# axs[0].set_ylabel('Density')
# axs[0].legend(title='True Class')

# # Plot Probability Density for Different Predicted Classes
# for class_label in df['class'].unique():
#     sns.kdeplot(df[df['class'] == class_label]['pos'], label=class_label, ax=axs[1])

# axs[1].set_title('Probability Density for Different Predicted Classes')
# axs[1].set_xlabel('Probability')
# axs[1].set_ylabel('Density')
# axs[1].legend(title='Predicted Class')

# plt.tight_layout()
# plt.show()

### Non-Linear SVM

from sklearn.kernel_approximation import Nystroem
from sklearn.linear_model import SGDClassifier
from sklearn.preprocessing import StandardScaler

# Standardize the features to have mean=0 and variance=1
scaler = StandardScaler(with_mean=False).fit(X_train_res)
X_train_scaled = scaler.transform(X_train_res)
X_train_sample_scaled = scaler.transform(X_train_sample)
X_dev_scaled = scaler.transform(X_dev)

# Use the Nystroem method to approximate a radial basis function (RBF) kernel
nystroem = Nystroem(kernel='rbf', gamma=0.2, random_state=42, n_components=1500)
nystroem.fit(X_train_scaled)
X_train_transformed = nystroem.transform(X_train_scaled)
X_train_sample_transformed = nystroem.transform(X_train_sample_scaled)
X_dev_transformed = nystroem.transform(X_dev_scaled)

# Initialize a LogisticRegression model, ‘log_loss’ gives logistic regression, a probabilistic classifier.
clf= SGDClassifier(loss='hinge',random_state=42, alpha=1e-4, n_iter_no_change=3, early_stopping=False)

# The partial fit if you ran out of RAM
batch_size = 50000
n_batches = X_train_transformed.shape[0] // batch_size

# partial fitting
for i in tqdm(range(n_batches)):
    start = i * batch_size
    end = (i + 1) * batch_size
    batch_X = X_train_transformed[start:end]
    batch_y = y_train_res[start:end]
    if i == 0:
        clf.partial_fit(batch_X, batch_y, classes=np.unique(y_train_res)) 
        # The `.A` here converts the sparse matrix to a dense matrix.
        # This is necessary because GaussianNB doesn't support sparse matrices.
        # We also specify the classes parameter in the first call to partial_fit.
    else:
        clf.partial_fit(batch_X, batch_y)

# get the results
y_pred = clf.predict(X_train_sample_transformed)
# y_pred_proba = clf.predict_proba(X_train_sample.toarray())

# Convert lists to arrays for further use
y_pred = np.array(y_pred)
# y_pred_proba = np.array(y_pred_proba)

print("Train")
print("accuracy_score",accuracy_score(y_train_sample, y_pred))
print("balanced_accuracy_score",balanced_accuracy_score(y_train_sample, y_pred))
print("average_precision_score",average_precision_score(y_train_sample, y_pred))
print(classification_report(y_train_sample, y_pred))

# get the results
y_pred = clf.predict(X_dev_transformed)
# y_pred_proba = clf.predict_proba(X_dev.toarray())

# Convert lists to arrays for further use
y_pred = np.array(y_pred)
# y_pred_proba = np.array(y_pred_proba)

print("Train")
print("accuracy_score",accuracy_score(y_dev, y_pred))
print("balanced_accuracy_score",balanced_accuracy_score(y_dev, y_pred))
print("average_precision_score",average_precision_score(y_dev, y_pred))
print(classification_report(y_dev, y_pred))
